pyproject.toml
[tool.poetry]
name = "zandbak"
version = "0.1.0"
description = ""
authors = ["GBPeters <gijstebrij@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
pandas = "^2.0"
snowflake-connector-python = {extras = ["pandas"], version = "^3.5.0"}
sqlalchemy = "^1"
snowflake-sqlalchemy = "^1.5.1"
python-dotenv = "^1.0.0"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
